requirements.txt
python-dotenv
--pre
openai
